refactor(single-day-view): started refactoring of table to distinct component
#41
