+++ src/components/SingleDayTimeEntryList.tsx
@@ -1,13 +1,11 @@
-import { Input, Table, Tag } from "antd"
+import format from "date-fns/format"
 import { useStoreState } from "pullstate"
 import { useQuery } from "react-query"
-import { formatDuration, formatTime } from "../services/date"
 import { TogglService } from "../services/toggl/TogglService"
 import { ProjectStore, SettingsStore, SingleDayViewStore } from "../store"
 import { TimeEntry } from "../types"
 import { DaySelector } from "./DaySelector"
-
-const { Column } = Table
+import { TimeEntryList } from "./TimeEntryList"
 
 // FIXME write some tests
 export function SingleDayTimeEntryList() {
@@ -18,51 +16,18 @@ export function SingleDayTimeEntryList() {
 
   // const { status, data, error } = useQuery<TimeEntry[], Error>(
   const { data } = useQuery<TimeEntry[], Error>(
-    ["todaysTimeEntries", day],
+    ["todaysTimeEntries", format(day, "yyyy-MM-dd")],
     async () => {
       return TogglService.getInstance(token).fetchTimeEntriesOfDay(day)
     },
     { enabled: projects.projects.length > 0 && !!token, retry: 0 }
   )
 
-  function renderProject(record: TimeEntry) {
-    return (
-      <Tag color={record.project.color}>
-        {record.project.name} | {record.project.client.name}
-      </Tag>
-    )
-  }
-
-  function renderStartStop(record: TimeEntry) {
-    const stop = "stop" in record ? formatTime(record.stop!) : ""
-    return (
-      <div className="wrap-start-end">
-        <Input defaultValue={formatTime(record.start)} />
-        <Input defaultValue={stop} />
-      </div>
-    )
-  }
-
-  function renderSum(record: TimeEntry) {
-    return <>{formatDuration(record.duration)}</>
-  }
-
-  function renderActions() {
-    return <>Actions</>
-  }
-
   // FIXME wrap the table into a form, render start+end as inputs already
-  // FIXME move table to own component incl. renderers to enable for better testing
   return (
     <>
       <DaySelector />
-      <Table dataSource={data} rowKey="id" showHeader={false} pagination={false}>
-        <Column title="Description" dataIndex="description" key="description" width="42%" />
-        <Column title="Project/Client" key="project" render={renderProject} width="30%" />
-        <Column title="Start/Stop" key="startstop" render={renderStartStop} width="12%" />
-        <Column title="Sum" key="sum" render={renderSum} width="8%" />
-        <Column title="Actions" key="actions" render={renderActions} width="8%" />
-      </Table>
+      <TimeEntryList entries={data} />
     </>
   )
 }
